requirements.txt
python-telegram-bot~=21.4
SQLAlchemy~=2.0.32
typing_extensions~=4.12.2
toml~=0.10.2
pex~=2.17.0
aiosqlite~=0.20.0
